+++ lib/payload/globals/AboutPage.ts
@@ -1,6 +1,7 @@
 import type { GlobalConfig } from "payload";
 import { triggerDeployHook } from "../deploy-hook";
 import { createMetadataFields } from "./fields/metadata-fields";
+import { createPageIntroFields } from "./fields/page-intro-fields";
 
 export const AboutPage: GlobalConfig = {
 	slug: "about-page",
@@ -15,16 +16,5 @@ export const AboutPage: GlobalConfig = {
 		read: () => true,
 		update: ({ req }) => !!req.user,
 	},
-	fields: [
-		{
-			name: "pageTitle",
-			type: "text",
-			required: true,
-			label: "Page Title",
-			admin: {
-				description: "The heading displayed on the page (H1)",
-			},
-		},
-		createMetadataFields(),
-	],
+	fields: [createPageIntroFields(), createMetadataFields()],
 };
